fix column data type

+++ database/migrations/2021_12_20_132138_create_staking_option_table.php
@@ -17,7 +17,7 @@ class CreateStakingOptionTable extends Migration
         Schema::create('staking_option', function (Blueprint $table) {
             $table->tinyIncrements('id');
             $table->string('vesting_period');
-            $table->unsignedTinyInteger('bonus_percentage');
+            $table->unsignedSmallInteger('bonus_percentage');
             $table->timestamps();
         });
